services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.7",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-jose>=3.5.0",
    "structlog>=25.5.0",
    "supabase==2.*",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.44.0",
    "anvx-core",
]

[tool.uv.sources]
anvx-core = { workspace = true }

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]
